+++ test/hamlit/optimization_test.rb
@@ -31,5 +31,10 @@ describe 'optimization' do
       haml = %q|%span= "jruby#{9000}#{dynamic}"|
       assert_equal true, compiled_code(haml).include?(%|<span>jruby9000|)
     end
+
+    it 'detects a static part recursively' do
+      haml = %q|%input{ value: "#{ "hello#{ hello }" }" }|
+      assert_equal true, compiled_code(haml).include?(%|value='hello|)
+    end
   end
 end